+++ lint/src/main/java/permissions/dispatcher/PermissionsDispatcherIssueRegistry.java
@@ -13,6 +13,12 @@ public final class PermissionsDispatcherIssueRegistry extends IssueRegistry {
         return Arrays.asList(
                 CallNeedsPermissionDetector.ISSUE,
                 CallOnRequestPermissionsResultDetector.ISSUE,
-                NoCorrespondingNeedsPermissionDetector.ISSUE);
+                NoCorrespondingNeedsPermissionDetector.ISSUE,
+                NoDelegateOnResumeDetector.ISSUE);
+    }
+
+    @Override
+    public int getApi() {
+        return com.android.tools.lint.detector.api.ApiKt.CURRENT_API;
     }
 }