+++ processor/src/main/kotlin/permissions/dispatcher/processor/impl/java/JavaSupportFragmentProcessorUnit.kt
@@ -2,19 +2,22 @@ package permissions.dispatcher.processor.impl.java
 
 import com.squareup.javapoet.MethodSpec
 import permissions.dispatcher.processor.util.*
+import javax.annotation.processing.Messager
 import javax.lang.model.type.TypeMirror
 
 /**
  * ProcessorUnit implementation for Fragments defined in the support-v4 library.
  */
-class JavaSupportFragmentProcessorUnit : JavaBaseProcessorUnit() {
+class JavaSupportFragmentProcessorUnit(messager: Messager) : JavaBaseProcessorUnit(messager) {
+
+    override fun isDeprecated(): Boolean = false
 
     override fun getTargetType(): TypeMirror {
         return typeMirrorOf("android.support.v4.app.Fragment")
     }
 
     override fun getActivityName(targetParam: String): String {
-        return targetParam + ".getActivity()"
+        return "$targetParam.getActivity()"
     }
 
     override fun addShouldShowRequestPermissionRationaleCondition(builder: MethodSpec.Builder, targetParam: String, permissionField: String, isPositiveCondition: Boolean) {
@@ -24,5 +27,4 @@ class JavaSupportFragmentProcessorUnit : JavaBaseProcessorUnit() {
     override fun addRequestPermissionsStatement(builder: MethodSpec.Builder, targetParam: String, permissionField: String, requestCodeField: String) {
         builder.addStatement("\$N.requestPermissions(\$N, \$N)", targetParam, permissionField, requestCodeField)
     }
-
 }