+++ processor/src/main/kotlin/permissions/dispatcher/processor/impl/kotlin/KotlinNativeFragmentProcessorUnit.kt
@@ -1,14 +1,17 @@
 package permissions.dispatcher.processor.impl.kotlin
 
 import com.squareup.kotlinpoet.FunSpec
-import permissions.dispatcher.processor.util.*
+import permissions.dispatcher.processor.util.typeMirrorOf
+import javax.annotation.processing.Messager
 import javax.lang.model.type.TypeMirror
 
-class KotlinNativeFragmentProcessorUnit : KotlinBaseProcessorUnit() {
+class KotlinNativeFragmentProcessorUnit(messager: Messager) : KotlinBaseProcessorUnit(messager) {
+
+    override fun isDeprecated(): Boolean = true
 
     override fun getTargetType(): TypeMirror = typeMirrorOf("android.app.Fragment")
 
-    override fun getActivityName(): String = "activity"
+    override fun getActivityName(targetParam: String): String = "$targetParam.activity"
 
     override fun addShouldShowRequestPermissionRationaleCondition(builder: FunSpec.Builder, permissionField: String, isPositiveCondition: Boolean) {
         val condition = if (isPositiveCondition) "" else "!"